% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{catConstants}
\alias{catConstants}
\title{Display the current constants}
\usage{
catConstants()
}
\description{
Prints the constants currently associated with reliability assessment:

* ramp load rate (k);

* sample size of reliability simulation for each \eqn{\theta};

* start time;

* return period (in hours);

* time step of ODE solver.
}
