% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{constLoad_4500_1Y}
\alias{constLoad_4500_1Y}
\title{Example data with \eqn{\tau_c = 4500} and \eqn{t_c = 1Y (8760 hours)}}
\format{A vector with 300 entries}
\usage{
constLoad_4500_1Y
}
\description{
An example dataset containing 300 failure times which can be used to obtain posterior samples of
\deqn{\theta = (\mu_a, \sigma_a, \mu_b, \sigma_b, \mu_c, \sigma_c, \mu_n, \sigma_n, \mu_{\sigma_0} , \sigma_{\sigma_0})}{\theta = (\mu_a, \sigma_a, \mu_b, \sigma_b, \mu_c, \sigma_c, \mu_n, \sigma_n, \mu_\sigma0, \sigma_\sigma0)}
 via the \code{\link{abcMCMC}} algorithm.
}
\keyword{datasets}
