% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\docType{data}
\name{default_param}
\alias{default_param}
\title{Default \eqn{\theta}}
\format{An object of class \code{numeric} of length 10.}
\usage{
default_param
}
\description{
Contains the values of \eqn{\theta = (\mu_a, \sigma_a, \mu_b, \sigma_b, \mu_c, \sigma_c, \mu_n, \sigma_n, \mu_{\sigma_0} , \sigma_{\sigma_0})}{\theta = (\mu_a, \sigma_a, \mu_b, \sigma_b, \mu_c, \sigma_c, \mu_n, \sigma_n, \mu_\sigma0 , \sigma_\sigma0)}.
}
\details{
Example values of \eqn{\theta} from preliminary data analysis of Hemlock load duration data.
}
\examples{
default_param
#modify one of the values
default_param[10] = 0.049


}
\references{
Yang, C. H., Zidek, J. V., & Wong, S. W. (2019). Bayesian analysis of accumulated damage models in lumber reliability. Technometrics, 61(2), 233-245.
}
\keyword{datasets}
