% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{print_LoadProfile}
\alias{print_LoadProfile}
\title{Print the load profile parameters}
\usage{
print_LoadProfile(paramName)
}
\arguments{
\item{paramName}{the name of the parameter in load profile}
}
\description{
Displays the value of parameters in the current load profile.
}
