% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_Const.R
\name{set_returnPeriod}
\alias{set_returnPeriod}
\title{Set end time of reliability assessment}
\usage{
set_returnPeriod(time_end, verbose = FALSE)
}
\arguments{
\item{time_end}{end of simulation period (in hours)}

\item{verbose}{prints information message to the console if TRUE}
}
\description{
Modifies the ending time of the simulation period (in hours) for the reliability assessment.
In \code{\link{solveCanADM}}, the load profile is simulated for a return period of \code{time_end} to determine whether failure has occurred.
}
\details{
The default is a reliability assessment period of 262980 hours (30 years).
}
\examples{
set_returnPeriod(131400)


}
