% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_Const.R
\name{set_simuSampleSize}
\alias{set_simuSampleSize}
\title{Set the number of simulated samples}
\usage{
set_simuSampleSize(n, verbose = FALSE)
}
\arguments{
\item{n}{the number of samples to simulate for each \eqn{\theta} to estimate failure probability}

\item{verbose}{prints information message to the console if TRUE}
}
\description{
Modifies the sample size of reliability simulations for \code{\link{solveCanADM}}.
}
\details{
The default number of samples is 1000
}
\examples{
set_simuSampleSize(10000)


}
