% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_Const.R
\name{set_timeStep}
\alias{set_timeStep}
\title{Set time step}
\usage{
set_timeStep(time_step, verbose = FALSE)
}
\arguments{
\item{time_step}{step of time}

\item{verbose}{prints information message to the console if TRUE}
}
\description{
Modifies time step (in hours).
The step size to use when solving the Canadian model numerically with \code{\link{solveCanADM}} for reliability assessments.
}
\details{
The default step of time is 100 hours.
}
\examples{
set_timeStep(50)


}
